backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
SQLAlchemy==2.0.31
python-jose[cryptography]==3.3.0
beautifulsoup4==4.12.3
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
pytest==8.2.2
httpx==0.27.0

